Update logits_chunk dtype handling for numerical stability
* Change logits_chunk to use float() instead of _input_chunk.dtype
* Comment out the line that switches logits_chunk back to the original dtype

+++ src/liger_kernel/ops/fused_linear_cross_entropy.py
@@ -70,7 +70,7 @@ def fused_linear_cross_entropy_forward(
         n_non_ignore = (target_chunk != ignore_index).sum().item()
 
         # when doing CE, use the upcasted precision
-        logits_chunk = logits_chunk.to(_input_chunk.dtype)
+        logits_chunk = logits_chunk.float()
 
         # ensure _input and target are contiguous
         logits_chunk = logits_chunk.contiguous()
@@ -98,7 +98,7 @@ def fused_linear_cross_entropy_forward(
         # w.r.t. logits in fp32 for numerical stability especially as the num classes (vocab size) is huge.
         # (reference: https://github.com/huggingface/transformers/blob/v4.42.4/src/transformers/models/llama/modeling_llama.py#L1194)
         # Propagating to lm_head's backward, we'll switch back to the original dtype.
-        logits_chunk = logits_chunk.to(dtype)
+        # logits_chunk = logits_chunk.to(dtype)
 
         # gradient of logits_chunk is computed in-place by the above triton kernel and is of shape: chunk_size x V
         # thus grad_input[start_idx: end_idx] should be of shape: chunk_size x H
